Make notifications length label vertically-centered
Previously, it was only centered when the Gnome Shell theme is the
default. Change it to cover other themes as well (e.g. Sweet Dark
for my case).

Although there is a tiny spacing error in the default Gnome theme,
but that should not matter much, because changing it would affect
other themes, and it IS centered, but not strictly in align with
the centered Github icon (maybe a fix should be done in the theme
itself?).